backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
redis>=5.0.0
celery>=5.3.0
python-multipart>=0.0.6
jinja2>=3.1.0
aiofiles>=23.0.0
httpx>=0.25.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
slowapi>=0.1.9
python-magic>=0.4.27
psutil>=5.9.0
pydantic-settings>=2.0.0

# audio_merge 모듈 의존성
pydub>=0.25.1

# Code quality tools  
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
